requirements.txt
flask==1.0.2
#gunicorn==19.8.1
numpy==1.14.1
matplotlib==2.2.2
postgres==2.2.1
progress==1.4.0
pytest==3.6.2
PyYAML==4.2b1
schedule==0.5.0
tensorflow==1.8.0
#https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow-1.10.1-cp36-cp36m-linux_x86_64.whl
